refactor: again remove singleton GameEngine and just keep a global instance of EnTT Registry

+++ include/player.h
@@ -21,7 +21,7 @@ namespace mitmeo
         public:
             Player();
 
-            uint16_t fire_interval;
+            uint32_t fire_interval;
 
             void set_fire_interval(uint8_t ms);
             void update(uint32_t time_ms);
